src/requirements.txt
altair==5.0.1
annotated-types==0.5.0
attrs==23.1.0
blinker==1.6.2
blis==0.7.10
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
confection==0.1.1
cymem==2.0.7
dnspython==2.4.2
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.6.0/en_core_web_md-3.6.0-py3-none-any.whl#sha256=1618fc62e7aedc2b778caa3ea6e9c8280beac3b83fe7b93316cbb9e8f3abe7c5
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
langcodes==3.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
murmurhash==1.0.9
numpy==1.25.2
packaging==23.1
pandas==2.0.3
pathy==0.10.2
Pillow==9.5.0
preshed==3.0.8
protobuf==4.24.0
pyarrow==12.0.1
pydantic==2.1.1
pydantic_core==2.4.0
pydeck==0.8.0
Pygments==2.16.1
pymongo==3.11.0
Pympler==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
streamlit==1.25.0
tenacity==8.2.3
thinc==8.1.12
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.21.2
wasabi==1.1.2
watchdog==3.0.0
zipp==3.16.2